requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.6
Django==4.2.4
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2==2.9.7
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
ruff==0.0.284
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
